Created a generic "world" structure to hold all the objects that can be rendered (HittableList).
Added a trait for "hittable objects" (Hittable).  Not sure if we'll use this in the future, as it's made redundant by the enum and static method calls in HittableList.

Moved sphere into its own struct, that can be hit-tested against.

Reworked main loop to use this object database and generic hit testing function.